requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
allure-pytest==2.14.0
allure-python-commons==2.14.0
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
blinker==1.8.2
bs4==0.0.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
cryptography==44.0.2
Cython==3.0.12
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
finlab==1.2.21
FinMind==1.7.8
firebase-admin==6.7.0
Flask==3.0.3
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
loguru==0.7.3
lxml==5.3.1
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
odfpy==1.4.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pluggy==1.5.0
prettytable==3.11.0
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.4
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyecharts==2.0.8
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2025.1
requests==2.32.3
rsa==4.9
selenium==4.27.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
ta==0.5.25
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.12.2
twstock==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2
gunicorn==21.2.0
